refactor(app): extract logout lifecycle
