+++ src/app/inputs.rs
@@ -29,7 +29,7 @@ impl App<'_> {
                     // further input until Event::LogoutResult resolves.
                     return;
                 }
-                self.resolve_logout_confirmation(key);
+                self.handle_logout_input(key);
                 return;
             }
 
@@ -380,69 +380,6 @@ impl App<'_> {
         self.action_notice = Some(crate::app::actions::ActionNotice::Unavailable(
             action.into(),
         ));
-    }
-
-    fn begin_logout_confirmation(&mut self) {
-        self.pending_logout = true;
-        self.logout_menu_index = 0;
-    }
-
-    /// Resolve the logout confirmation as a top-right menu, matching the
-    /// message-menu interaction: j/k (or Up/Down) move the selection, Enter
-    /// confirms it, Esc/n cancel. y is kept as a fast "confirm" and N as a
-    /// fast "cancel".
-    fn resolve_logout_confirmation(&mut self, key: Key) {
-        match key.code {
-            KeyCode::Char('y') | KeyCode::Char('Y') => {
-                self.pending_logout = false;
-                self.confirm_logout();
-            }
-            KeyCode::Char('n') | KeyCode::Char('N') | KeyCode::Esc => {
-                self.pending_logout = false;
-                self.logout_menu_index = 0;
-                self.action_notice = Some(crate::app::actions::ActionNotice::Cancelled);
-            }
-            KeyCode::Enter => {
-                if self.logout_menu_index == 0 {
-                    self.pending_logout = false;
-                    self.confirm_logout();
-                } else {
-                    self.pending_logout = false;
-                    self.logout_menu_index = 0;
-                    self.action_notice = Some(crate::app::actions::ActionNotice::Cancelled);
-                }
-            }
-            KeyCode::Char('j') | KeyCode::Down => {
-                self.logout_menu_index = (self.logout_menu_index + 1) % 2;
-            }
-            KeyCode::Char('k') | KeyCode::Up => {
-                self.logout_menu_index = (self.logout_menu_index + 1) % 2;
-            }
-            _ => {}
-        }
-    }
-
-    fn confirm_logout(&mut self) {
-        // The bridge runs the remove-companion-device IQ asynchronously so
-        // the event loop stays responsive. Keep intercepting input and let
-        // Event::LogoutResult drive the local cleanup / quit / error path.
-        self.pending_logout = true;
-        self.logout_in_progress = true;
-        wr::logout();
-    }
-
-    pub fn finish_logout(&mut self) {
-        self.pending_logout = false;
-        self.logout_in_progress = false;
-        self.logout_menu_index = 0;
-        // Stop the history writer so the sqlite files are released, then wipe
-        // BOTH databases plus media: a logout is a full account sign-out, and
-        // pairing a different account must never see the previous one's data.
-        self.db_handler.stop();
-        wipe_sqlite_file(&self.whatsmeow_db);
-        wipe_sqlite_file(&self.whatsmeow_db.with_file_name("whatsapp.db"));
-        clear_media_dir(&self.media_path);
-        self.should_quit = true;
     }
 
     fn selected_message_is_deleted(&self) -> bool {
@@ -1162,12 +1099,7 @@ impl App<'_> {
     fn select_next(&mut self) {
         match self.focus_pane {
             FocusPane::SectionRail => {
-                if self.rail_on_logout {
-                    self.rail_on_logout = false;
-                    self.selected_section = Section::Chats;
-                } else if self.selected_section == Section::Communities {
-                    self.rail_on_logout = true;
-                } else {
+                if !self.move_logout_selection_next() {
                     self.selected_section = self.selected_section.next();
                 }
             }
@@ -1199,12 +1131,7 @@ impl App<'_> {
     fn select_previous(&mut self) {
         match self.focus_pane {
             FocusPane::SectionRail => {
-                if self.rail_on_logout {
-                    self.rail_on_logout = false;
-                    self.selected_section = Section::Communities;
-                } else if self.selected_section == Section::Chats {
-                    self.rail_on_logout = true;
-                } else {
+                if !self.move_logout_selection_previous() {
                     self.selected_section = self.selected_section.previous();
                 }
             }
@@ -1236,8 +1163,7 @@ impl App<'_> {
     fn jump_top(&mut self) {
         match self.focus_pane {
             FocusPane::SectionRail => {
-                self.selected_section = crate::app::actions::Section::Chats;
-                self.rail_on_logout = false;
+                self.jump_logout_selection_top();
             }
             FocusPane::ChatList => {
                 if self.selected_section == Section::Status {
@@ -1264,8 +1190,7 @@ impl App<'_> {
     fn jump_bottom(&mut self) {
         match self.focus_pane {
             FocusPane::SectionRail => {
-                self.selected_section = crate::app::actions::Section::Communities;
-                self.rail_on_logout = true;
+                self.jump_logout_selection_bottom();
             }
             FocusPane::ChatList => {
                 if self.selected_section == Section::Status {
@@ -1539,41 +1464,6 @@ fn message_urls(message: &wr::Message) -> Vec<String> {
     };
     text.map(crate::url::extract_openable_urls)
         .unwrap_or_default()
-}
-
-/// Remove a sqlite database together with its `-wal` and `-shm` sidecars.
-/// Used during logout to guarantee a full account sign-out.
-fn wipe_sqlite_file(path: &Path) {
-    for suffix in [
-        std::ffi::OsStr::new(""),
-        std::ffi::OsStr::new("-wal"),
-        std::ffi::OsStr::new("-shm"),
-    ] {
-        let mut os = path.as_os_str().to_os_string();
-        os.push(suffix);
-        let file = std::path::PathBuf::from(&os);
-        match std::fs::remove_file(&file) {
-            Ok(()) => {}
-            Err(err) if err.kind() == std::io::ErrorKind::NotFound => {}
-            Err(err) => log::warn!("Failed to remove {}: {err}", file.display()),
-        }
-    }
-}
-
-/// Remove every entry under the media cache directory. A missing directory is
-/// fine; a nested directory tree is removed recursively.
-fn clear_media_dir(media_path: &Path) {
-    let Ok(entries) = std::fs::read_dir(media_path) else {
-        return;
-    };
-    for entry in entries.flatten() {
-        let path = entry.path();
-        if path.is_dir() {
-            let _ = std::fs::remove_dir_all(&path);
-        } else {
-            let _ = std::fs::remove_file(&path);
-        }
-    }
 }
 
 pub fn composer_action_for_editing_key(key: &Key) -> ComposerAction {
